+++ ZXSinclair.Net/Hardware/MemoryBuffer.cs
@@ -23,7 +23,7 @@ namespace ZXSinclair.Net.Hardware
     {
         private bool disposedValue;
 
-        public MemoryBuffer(int size = 0, byte* ptr = null, UInt32 offset = 0)
+        public MemoryBuffer(int size = 0, byte* ptr = null, ushort offset = 0)
         {
             if (size < 0)
                 throw new InvalidOperationException();
@@ -49,12 +49,12 @@ namespace ZXSinclair.Net.Hardware
             }
         }
 
-        public UInt32 Offset { get; }
+        public ushort Offset { get; }
         public int Size { get; }
         public IntPtr Buffer { get; private set; }
         public byte* BufferPtr { get; protected set; }
 
-        public MemoryBuffer New(UInt32 offset, int size)
+        public MemoryBuffer New(ushort offset, int size)
         {
             if (offset > Size || offset + size > Size)
                 new MemoryBuffer(0);
@@ -104,7 +104,7 @@ namespace ZXSinclair.Net.Hardware
         {
         }
 
-        public byte Read(UInt32 address)
+        public byte Read(ushort address)
         {
             if (address >= Size)
                 return 0;
@@ -112,7 +112,7 @@ namespace ZXSinclair.Net.Hardware
             return *(BufferPtr + address);
         }
 
-        public void Write(UInt32 address, byte data)
+        public void Write(ushort address, byte data)
         {
             if (address >= Size)
                 return;
